refactor: move go/web to go/common/web
Another step toward simpler repo layout. This is an exception to the guideline that code with no dependency on multigres types should live in go/tools. These templates do not (yet) depend on anything else in multigres, but they still feel very multigres-specific rather than being a generic tool
