+++ Makefile
@@ -66,7 +66,7 @@ generate: parser
 # Build Go binaries only
 build:
 	mkdir -p bin/
-	cp external/pico/pico.* go/web/templates/css/
+	cp external/pico/pico.* go/common/web/templates/css/
 	go build -o bin/multigateway ./go/cmd/multigateway
 	go build -o bin/multipooler ./go/cmd/multipooler
 	go build -o bin/pgctld ./go/cmd/pgctld
@@ -79,7 +79,7 @@ build-all: proto parser build
 
 # Clean build artifacts
 clean:
-	rm -f go/web/templates/css/pico.*
+	rm -f go/common/web/templates/css/pico.*
 	go clean -i ./go/...
 	rm -f bin/*
