+++ go/common/web/templates/favicon.ico
