+++ go/common/web/templates/gateway_index.html
