+++ go/common/web/templates/status_text.tmpl
