+++ go/common/web/templates/timestamp.tmpl
